mojo(wasm): transport process-local shared memory regions
Move Wasm shared-memory capabilities through one-shot Mojo platform-handle tokens. Preserve ownership across wrap, unwrap, validation failure, and duplicate-region paths while keeping native and cross-node transport explicitly unavailable.

Tests:
- third_party/depot_tools/gn check out/wasm //tools/wasm:m1_mojo_smoke
- third_party/depot_tools/ninja -C out/wasm -j2 m1_mojo_smoke
- PYTHONDONTWRITEBYTECODE=1 python3 tools/wasm/run_node_smoke.py --case mojo --timeout 30
- PYTHONDONTWRITEBYTECODE=1 python3 tools/wasm/run_browser_smoke.py --case mojo --browser /usr/bin/google-chrome-stable --out-dir out/wasm --timeout 30 --no-sandbox
- PYTHONDONTWRITEBYTECODE=1 python3 -m unittest tools.wasm.tests.test_m0_tools.NodeRunnerTest.test_mojo_case_requires_complete_result_contract tools.wasm.tests.test_m3_base_source_contract.M3BaseSourceContractTest.test_mojo_transfers_process_local_shared_memory_capabilities tools.wasm.tests.test_m3_base_source_contract.M3BaseSourceContractTest.test_mojo_platform_file_transport_is_explicitly_unsupported tools.wasm.tests.test_m3_ozone_source_contract.M3OzoneSourceContractTest.test_wasm_ui_fallbacks_are_explicit_and_non_native

Known limitations:
- Capabilities are valid only inside one WebAssembly application.
- Shared-buffer serialization across ipcz transport nodes remains unsupported.

+++ mojo/core/ipcz_driver/shared_buffer.h
@@ -41,7 +41,7 @@ class MOJO_SYSTEM_IMPL_EXPORT SharedBuffer : public Object<SharedBuffer> {
   // Helper to support the MojoWrapPlatformSharedMemoryRegion API.
   static scoped_refptr<SharedBuffer> CreateForMojoWrapper(
       base::span<const MojoPlatformHandle> mojo_platform_handles,
-      uint32_t size,
+      uint64_t size,
       const MojoSharedBufferGuid& mojo_guid,
       MojoPlatformSharedMemoryRegionAccessMode access);
